ModuleFieldPartitioner uses Long instead of Int

+++ kite-data/kite-data-core/src/main/java/org/kitesdk/data/spi/partition/ModuleFieldPartitioner.java
@@ -1,3 +1,18 @@
+/**
+ * Copyright 2013 Cloudera Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.kitesdk.data.spi.partition;
 
 import javax.annotation.Nullable;
@@ -7,7 +22,7 @@ import org.kitesdk.data.spi.FieldPartitioner;
 import com.google.common.base.Objects;
 import com.google.common.base.Predicate;
 
-public class ModuleFieldPartitioner extends FieldPartitioner<Integer, Integer> {
+public class ModuleFieldPartitioner extends FieldPartitioner<Long, Long> {
 
 	private int module;
 
@@ -16,28 +31,28 @@ public class ModuleFieldPartitioner extends FieldPartitioner<Integer, Integer> {
 	}
 
 	public ModuleFieldPartitioner(String sourceName, String name, int module) {
-		super(sourceName, name, Integer.class, Integer.class, module);
+		super(sourceName, name, Long.class, Long.class, module);
 
 		this.module = module;
 	}
 
 	@Override
-	public int compare(Integer o1, Integer o2) {
-		return Integer.compare(o1, o2);
+	public int compare(Long o1, Long o2) {
+		return Long.compare(o1, o2);
 	}
 
 	@Override
-	public Integer apply(Integer value) {
+	public Long apply(Long value) {
 		return value % module;
 	}
 
 	@Override
-	public Predicate<Integer> project(Predicate<Integer> predicate) {
+	public Predicate<Long> project(Predicate<Long> predicate) {
 		return null;
 	}
 
 	@Override
-	public Predicate<Integer> projectStrict(Predicate<Integer> predicate) {
+	public Predicate<Long> projectStrict(Predicate<Long> predicate) {
 		return null;
 	}
